+++ tests/integration/test_msgraph/conftest.py
@@ -3,11 +3,16 @@
 # This program is free software; you can redistribute it and/or modify it under the terms of GPLv2
 
 import pytest
+import subprocess
+import os
+from pathlib import Path
 
 from wazuh_testing.constants.paths.logs import WAZUH_LOG_PATH
+from wazuh_testing.constants.paths import WAZUH_PATH
 from wazuh_testing.modules.modulesd import patterns
 from wazuh_testing.tools.monitors.file_monitor import FileMonitor
 from wazuh_testing.utils import callbacks
+from wazuh_testing.utils.file import remove_file
 
 @pytest.fixture()
 def wait_for_msgraph_start():
@@ -17,3 +22,17 @@ def wait_for_msgraph_start():
                               'integration': 'ms-graph'
                           }))     
     assert (wazuh_log_monitor.callback_result == None), f'Error invalid configuration event not detected'
+
+@pytest.fixture(scope="session")
+def proxy_setup():
+    RESPONSES_PATH = Path(Path(__file__).parent, 'test_API', 'data', 'response_samples', 'responses.json')
+    m365proxy = subprocess.Popen(["/tmp/m365proxy/m365proxy", "--mocks-file", RESPONSES_PATH])
+    # Configurate proxy for Wazuh (will only work for systemctl start/restart)
+    subprocess.run("systemctl set-environment http_proxy=http://localhost:8000", shell=True)
+    remove_file(os.path.join(WAZUH_PATH, 'var', 'wodles', 'ms-graph-tenant_id-resource_name-resource_relationship'))
+
+    yield
+
+    subprocess.run("systemctl unset-environment http_proxy", shell=True)
+    m365proxy.kill()
+    m365proxy.wait()
